+++ src/Utilites/Ssr/Template/RecordList.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Omikron\FactFinder\Shopware6\Utilites\Ssr\Template;
 
+use Omikron\FactFinder\Shopware6\Utilites\Ssr\Exception\DetectRedirectCampaignException;
 use Omikron\FactFinder\Shopware6\Utilites\Ssr\SearchAdapter;
 use Symfony\Component\HttpFoundation\Request;
 
@@ -36,15 +37,41 @@ class RecordList
 
     /**
      * @SuppressWarnings(PHPMD.BooleanArgumentFlag)
+     *
+     * @throws DetectRedirectCampaignException
      */
     public function getContent(
         string $paramString,
         bool $isNavigationRequest = false
-    ) {
+    ): string {
+        $results = $this->searchResults($paramString, $isNavigationRequest);
+
+        // Support redirect campaigns for SSR
+        if ($this->getRedirectCampaign($results)) {
+            throw new DetectRedirectCampaignException($this->getRedirectCampaign($results));
+        }
+
+        return $this->renderResults($results, $paramString);
+    }
+
+    /**
+     * @SuppressWarnings(PHPMD.BooleanArgumentFlag)
+     */
+    public function searchResults(
+        string $paramString,
+        bool $isNavigationRequest = false
+    ): array {
         $paramString = strpos($paramString, 'p=') === 0
             ? sprintf('page=%s', substr($paramString, 2))
             : str_replace('&p=', '&page=', $paramString);
-        $results        = $this->searchAdapter->search($paramString, $isNavigationRequest, $this->salesChannelId);
+
+        return $this->searchAdapter->search($paramString, $isNavigationRequest, $this->salesChannelId);
+    }
+
+    public function renderResults(
+        array $results,
+        string $paramString
+    ): string {
         $records        = $results['records'] ?? [];
         $recordsContent = array_reduce(
             $records,
@@ -92,5 +119,16 @@ class RecordList
         preg_match(self::RECORD_PATTERN, $this->content, $match);
 
         $this->template = $match[0] ?? '';
+    }
+
+    private function getRedirectCampaign(array $result): ?string
+    {
+        if (!empty($result['campaigns'])) {
+            $campaign = array_search('REDIRECT', array_column($result['campaigns'], 'flavour'));
+
+            return $result['campaigns'][$campaign]['target']['destination'] ?? null;
+        }
+
+        return null;
     }
 }